+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fill `in_reply_to_account_id` value on `Status.reblog` object.
+- Fixed extraction of media IDs from urlencoded form data.
 
 ## [2.6.0] - 2023-12-25